requirements.txt
boto3
snowflake-connector-python
pandas
matplotlib